+++ CHANGELOG.md
@@ -2,3 +2,4 @@
 - added NBT materials
 - added command to get current material
 - added better for tool-detection on forge
+- fixed enchantment compat with Apotheosis and Zenith